+++ src/App.tsx
@@ -1,3 +1,4 @@
+import { useEffect, useState } from 'react'
 import { Navigate, Route, Routes } from 'react-router-dom'
 import './App.css'
 import { ContactsPage } from './Pages/contactsPage/ContactsPage'
@@ -6,14 +7,22 @@ import { HomePage } from './Pages/homePage/HomePage'
 import { PricePage } from './Pages/pricePage/PricePage'
 import { TeamPage } from './Pages/teamPage/TeamPage'
 import { Footer } from './layout/footer/Footer'
-import { Header } from './layout/header/Header'
-import { HeaderMobile } from './layout/header/headerMobile/HeaderMobile'
+import { DesktopHeader } from './layout/header/desktopHeader/DesktopHeader'
+import { MobileHeader } from './layout/header/mobileHeader/MobileHeader'
 
 function App() {
+	const [width, setWidth] = useState(window.innerWidth)
+	const breakpoint = 768
+
+	useEffect(() => {
+		const handleWindowResize = () => setWidth(window.innerWidth)
+		window.addEventListener('resize', handleWindowResize)
+		return () => window.removeEventListener('resize', handleWindowResize)
+	}, [])
+
 	return (
 		<div className='App'>
-			<Header />
-			<HeaderMobile />
+			{width < breakpoint ? <MobileHeader /> : <DesktopHeader />}
 			<Routes>
 				<Route path='/' element={<Navigate to={'/home'} />} />
 				<Route path='/home' element={<HomePage />} />
