+++ src/layout/header/mobileHeader/MobileHeader.tsx
@@ -1,12 +1,11 @@
 import { useState } from 'react'
-import styled from 'styled-components'
 import { Container } from '../../../components/Container'
 import { FlexWrapper } from '../../../components/FlexWrapper'
 import { Burger } from '../../../components/burger/Burger'
 import { Icon } from '../../../components/icon/Icon'
 import { Logo } from '../../../components/logo/Logo'
 import { Menu } from '../../../components/menu/Menu'
-import { theme } from '../../../styles/Theme'
+import { S } from './MobileHeader_Styles'
 
 export interface INavItem {
 	id: string
@@ -23,11 +22,11 @@ const NavDate: INavItem[] = [
 	{ id: 'RightNav3', link: 'contacts', text: 'Контакты' },
 ]
 
-export function HeaderMobile() {
+export const MobileHeader: React.FC = () => {
 	const [modalIsOpen, setModalIsOpen] = useState(true)
 
 	return (
-		<StyledHeaderMobile>
+		<S.MobileHeader>
 			<Container>
 				<FlexWrapper justify='space-between' align='center'>
 					<Icon
@@ -45,23 +44,6 @@ export function HeaderMobile() {
 					<Burger isOpen={modalIsOpen} setIsOpen={setModalIsOpen} />
 				</FlexWrapper>
 			</Container>
-		</StyledHeaderMobile>
+		</S.MobileHeader>
 	)
 }
-
-const StyledHeaderMobile = styled.header`
-	display: none;
-	position: absolute;
-	top: 0;
-	left: 0;
-	right: 0;
-	z-index: 5;
-	${Container} {
-		border-bottom: 1px solid ${theme.colors.font};
-		padding: 40px 15px 30px 15px;
-	}
-
-	@media ${theme.media.tablet} {
-		display: block;
-	}
-`
